requirements.txt
# MUSE Language - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# AI Services
openai==1.12.0
tiktoken==0.5.2
langchain==0.1.5
langchain-openai==0.0.5

# Speech Processing
SpeechRecognition==3.10.1
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# NLP
spacy==3.7.2
nltk==3.8.1
language-tool-python==2.7.1

# Task Queue
celery==5.3.6

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
httpx==0.26.0
python-dotenv==1.0.1
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0
